Move old field infos format to backwards-codecs.
