+++ lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene60/Lucene60FieldInfosFormat.java
@@ -14,11 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.lucene.codecs.lucene60;
+package org.apache.lucene.backward_codecs.lucene60;
 
-import java.io.IOException;
-import java.util.Collections;
-import java.util.Map;
 import org.apache.lucene.codecs.CodecUtil;
 import org.apache.lucene.codecs.DocValuesFormat;
 import org.apache.lucene.codecs.FieldInfosFormat;
@@ -36,6 +33,10 @@ import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.IOContext;
 import org.apache.lucene.store.IndexInput;
 import org.apache.lucene.store.IndexOutput;
+
+import java.io.IOException;
+import java.util.Collections;
+import java.util.Map;
 
 /**
  * Lucene 6.0 Field Infos format.
@@ -311,6 +312,11 @@ public final class Lucene60FieldInfosFormat extends FieldInfosFormat {
     }
   }
 
+  /**
+   * Note: although this format is only used on older versions, we need to keep the write logic
+   * in addition to the read logic. It's possible for field infos on older segments to be written
+   * to, for example as part of in-place doc values updates.
+   */
   @Override
   public void write(
       Directory directory,
